+++ wxc_sdk/as_api.py
@@ -50,10 +50,10 @@ CALLING_DATA_TIMEOUT_PROTECTION = False
 
 
 __all__ = ['AsAccessCodesApi', 'AsAgentCallerIdApi', 'AsAnnouncementApi', 'AsApiChild', 'AsAppServicesApi',
-           'AsAttachmentActionsApi', 'AsAuthCodesApi', 'AsAutoAttendantApi', 'AsBargeApi', 'AsCQPolicyApi',
-           'AsCallInterceptApi', 'AsCallParkApi', 'AsCallPickupApi', 'AsCallQueueApi', 'AsCallRecordingApi',
-           'AsCallWaitingApi', 'AsCallerIdApi', 'AsCallingBehaviorApi', 'AsCallparkExtensionApi', 'AsCallsApi',
-           'AsDetailedCDRApi', 'AsDeviceSettingsJobsApi', 'AsDevicesApi', 'AsDialPlanApi', 'AsDndApi', 'AsEventsApi',
+           'AsAttachmentActionsApi', 'AsAutoAttendantApi', 'AsBargeApi', 'AsCQPolicyApi', 'AsCallInterceptApi',
+           'AsCallParkApi', 'AsCallPickupApi', 'AsCallQueueApi', 'AsCallRecordingApi', 'AsCallWaitingApi',
+           'AsCallerIdApi', 'AsCallingBehaviorApi', 'AsCallparkExtensionApi', 'AsCallsApi', 'AsDetailedCDRApi',
+           'AsDeviceSettingsJobsApi', 'AsDevicesApi', 'AsDialPlanApi', 'AsDndApi', 'AsEventsApi',
            'AsExecAssistantApi', 'AsForwardingApi', 'AsGroupsApi', 'AsHotelingApi', 'AsHuntGroupApi',
            'AsIncomingPermissionsApi', 'AsInternalDialingApi', 'AsJobsApi', 'AsLicensesApi', 'AsLocationInterceptApi',
            'AsLocationMoHApi', 'AsLocationNumbersApi', 'AsLocationVoicemailSettingsApi', 'AsLocationsApi',
@@ -5663,77 +5663,80 @@ class AsNumbersApi(AsPersonSettingsApiChild):
         await self.put(url=url, params=params, data=body)
 
 
-class AsAuthCodesApi(AsPersonSettingsApiChild):
+class AsAccessCodesApi(AsPersonSettingsApiChild):
     """
-    API for person's outgoing permission authorization codes
+    API for workspace's outgoing permission access codes
     """
-    feature = 'outgoingPermission/authorizationCodes'
+    feature = 'outgoingPermission/accessCodes'
 
-    async def read(self, person_id: str, org_id: str = None) -> list[AuthCode]:
+    async def read(self, workspace_id: str, org_id: str = None) -> list[AuthCode]:
         """
-        Retrieve Authorization codes for a Workspace.
+        Retrieve Access codes for a Workspace.
 
-        Authorization codes are used to bypass permissions.
+        Access codes are used to bypass permissions.
 
         This API requires a full or read-only administrator auth token with a scope of spark-admin:workspaces_read or
         a user auth token with spark:workspaces_read scope can be used to read workspace settings.
 
-        :param person_id: Unique identifier for the workspace.
-        :type person_id: str
-        :param org_id: Workspace is in this organization. Only admin users of another organization (such as partners)
-            may use this parameter as the default is the same organization as the token used to access API.
+        :param workspace_id: Unique identifier for the workspace.
+        :type workspace_id: str
+        :param org_id: ID of the organization within which the workspace resides. Only admin users of another
+            organization (such as partners) may use this parameter as the default is the same organization as the token
+            used to access API.
         :type org_id: str
-        :return: list of authorization codes
+        :return: list of access codes
         :rtype: list of :class:`AuthCode`
         """
-        url = self.f_ep(person_id=person_id)
+        url = self.f_ep(person_id=workspace_id)
         params = org_id and {'orgId': org_id} or None
         data = await self.get(url, params=params)
-        return parse_obj_as(list[AuthCode], data['authorizationCodes'])
+        return parse_obj_as(list[AuthCode], data['accessCodes'])
 
-    async def delete_codes(self, person_id: str, access_codes: list[Union[str, AuthCode]], org_id: str = None):
+    async def delete_codes(self, workspace_id: str, access_codes: list[Union[str, AuthCode]], org_id: str = None):
         """
-        Modify Authorization codes for a workspace.
+        Modify Access codes for a workspace.
 
-        Authorization codes are used to bypass permissions.
+        Access codes are used to bypass permissions.
 
         This API requires a full or user administrator auth token with the spark-admin:workspaces_write scope or a
         user auth token with spark:workspaces_write scope can be used to update workspace settings.
 
-        :param person_id: Unique identifier for the workspace.
-        :type person_id: str
+        :param workspace_id: Unique identifier for the workspace.
+        :type workspace_id: str
         :param access_codes: authorization codes to remove
         :type access_codes: list[str]
-        :param org_id: Workspace is in this organization. Only admin users of another organization (such as partners)
-            may use this parameter as the default is the same organization as the token used to access API.
+        :param org_id: ID of the organization within which the workspace resides. Only admin users of another
+            organization (such as partners) may use this parameter as the default is the same organization as the token
+            used to access API.
         :type org_id: str
         """
-        url = self.f_ep(person_id=person_id)
+        url = self.f_ep(person_id=workspace_id)
         params = org_id and {'orgId': org_id} or None
         body = {'deleteCodes': [ac.code if isinstance(ac, AuthCode) else ac
                                 for ac in access_codes]}
         await self.put(url, params=params, json=body)
 
-    async def create(self, person_id: str, code: str, description: str, org_id: str = None):
+    async def create(self, workspace_id: str, code: str, description: str, org_id: str = None):
         """
-        Modify Authorization codes for a workspace.
+        Create new Access codes for the given workspace.
 
-        Authorization codes are used to bypass permissions.
+        Access codes are used to bypass permissions.
 
         This API requires a full or user administrator auth token with the spark-admin:workspaces_write scope or a
         user auth token with spark:workspaces_write scope can be used to update workspace settings.
 
-        :param person_id: Unique identifier for the workspace.
-        :type person_id: str
-        :param code: Indicates an authorization code.
+        :param workspace_id: Unique identifier for the workspace.
+        :type workspace_id: str
+        :param code: Indicates an access code.
         :type code: str
-        :param description: Indicates the description of the authorization code.
+        :param description: Indicates the description of the access code.
         :type description: str
-        :param org_id: Workspace is in this organization. Only admin users of another organization (such as partners)
-            may use this parameter as the default is the same organization as the token used to access API.
+        :param org_id: ID of the organization within which the workspace resides. Only admin users of another
+            organization (such as partners) may use this parameter as the default is the same organization as the token
+            used to access API.
         :type org_id: str
         """
-        url = self.f_ep(person_id=person_id)
+        url = self.f_ep(person_id=workspace_id)
         params = org_id and {'orgId': org_id} or None
         body = {'code': code,
                 'description': description}
@@ -5804,7 +5807,7 @@ class AsOutgoingPermissionsApi(AsPersonSettingsApiChild):
     #: Only available for workspaces and locations
     transfer_numbers: AsTransferNumbersApi
     #: Only available for workspaces
-    auth_codes: AsAuthCodesApi
+    access_codes: AsAccessCodesApi
 
     feature = 'outgoingPermission'
 
@@ -5815,14 +5818,14 @@ class AsOutgoingPermissionsApi(AsPersonSettingsApiChild):
             # auto transfer numbers API seems to only exist for workspaces
             self.transfer_numbers = AsTransferNumbersApi(session=session,
                                                        workspaces=True)
-            self.auth_codes = AsAuthCodesApi(session=session, workspaces=True)
+            self.access_codes = AsAccessCodesApi(session=session, workspaces=True)
         elif locations:
             self.transfer_numbers = AsTransferNumbersApi(session=session,
                                                        locations=True)
-            self.auth_codes = None
+            self.access_codes = None
         else:
             self.transfer_numbers = None
-            self.auth_codes = None
+            self.access_codes = None
 
     async def read(self, person_id: str, org_id: str = None) -> OutgoingPermissions:
         """
@@ -7481,93 +7484,6 @@ class AsTeamsApi(AsApiChild, base='teams'):
         url = self.ep(f'{team_id}')
         await super().delete(url=url)
         return
-
-
-class AsAccessCodesApi(AsApiChild, base='telephony/config/locations'):
-    """
-    Access codes API
-    """
-
-    def _endpoint(self, *, location_id: str, path: str = None) -> str:
-        """
-        location specific feature endpoint like
-        /v1/telephony/config/locations/{locationId}/outgoingPermission/accessCodes}
-
-        :meta private:
-        :param location_id: Unique identifier for the location.
-        :type location_id: str
-        :param path: additional path
-        :type: path: str
-        :return: full endpoint
-        :rtype: str
-        """
-        path = path and f'/{path}' or ''
-        ep = self.session.ep(f'telephony/config/locations/{location_id}/outgoingPermission/accessCodes{path}')
-        return ep
-
-    async def read(self, location_id: str, org_id: str = None) -> list[AuthCode]:
-        """
-        Get Location Access Code
-
-        Retrieve access codes details for a customer location.
-
-        Use Access Codes to bypass the set permissions for all persons/workspaces at this location.
-
-        Retrieving access codes details requires a full, user or read-only administrator auth token with a scope of
-        spark-admin:telephony_config_read.
-
-
-        :param location_id: Retrieve access codes details for this location.
-        :type location_id: str
-        :param org_id: Retrieve access codes details for a customer location in this organization
-        :type org_id: str
-        :return: list of :class:`wxc_sdk.common.CallPark`
-        """
-        params = org_id and {'orgId': org_id} or None
-        url = self._endpoint(location_id=location_id)
-        data = await self.get(url, params=params)
-        return parse_obj_as(list[AuthCode], data['accessCodes'])
-
-    async def create(self, location_id: str, access_codes: list[AuthCode], org_id: str = None) -> list[AuthCode]:
-        """
-        Create access code in location
-
-        :param location_id: Add new access code for this location.
-        :type location_id: str
-        :param access_codes: Access code details
-        :type access_codes: list of :class:`wxc_sdk.common.AuthCode`
-        :param org_id: Add new access code for this organization.
-        :type org_id: str
-        """
-        params = org_id and {'orgId': org_id} or None
-        url = self._endpoint(location_id=location_id)
-        body = {'accessCodes': [json.loads(ac.json()) for ac in access_codes]}
-        await self.post(url, json=body, params=params)
-
-    async def delete_codes(self, location_id: str, access_codes: list[Union[str, AuthCode]],
-                     org_id: str = None) -> list[AuthCode]:
-        """
-        Delete Access Code Location
-
-        Deletes the access code details for a particular location for a customer.
-
-        Use Access Codes to bypass the set permissions for all persons/workspaces at this location.
-
-        Modifying the access code location details requires a full administrator auth token with a scope
-        of spark-admin:telephony_config_write.
-
-        :param location_id: Deletes the access code details for this location.
-        :type location_id: str
-        :param access_codes: access codes to delete
-        :type access_codes: list of :class:`wxc_sdk.common.AuthCode` or str
-        :param org_id: Delete access codes from this organization.
-        :type org_id: str
-        """
-        params = org_id and {'orgId': org_id} or None
-        url = self._endpoint(location_id=location_id)
-        body = {'deleteCodes': [ac.code if isinstance(ac, AuthCode) else ac
-                                for ac in access_codes]}
-        await self.put(url, json=body, params=params)
 
 
 class AsForwardingApi: